requirements/base.txt
coreapi==2.3.3
Django==3.0.8
djangorestframework==3.11.0
Markdown==3.2.2
